Show completed exercise
Finished the first exercise, with a few additions. I wanted to add a
font family property to each of the elements, in order to better match
the desired outcomes. After completing, I formatted the code to match
the 'solution' files. I had most of my code on one line for some
elements.

Additionally, I wanted to practice writing out a longer commit message.
I set up git to open the text editor when no message is included in a
commit. This looks like a good place to end the message.